Added AsyncTask to send relay value.
Added the functionality to set the relay value. This can be done by
pressing a button showing the relay name in the main screen. The button
should be presented as pressed if the value of the relay is 1 (set) or
as normal if the value is 0 (not set). This is to be fixed still!!